cambios diminutos a base de datos, controlador factura arreglado, detalle factura y gestion de ventas totalmente funcional y arreglo diseño

+++ Crud/controlador/controlador.producto.php
@@ -1,12 +1,30 @@
 <?php
-    require '../Dao/productoDao.php';
-    require '../Dto/productoDto.php';
-    require '../utilidades/conexion.php';
+require '../Dao/productoDao.php';
+require '../Dto/productoDto.php';
+session_start();
 
-if (isset($_POST['registrarProducto'])){
+if (!isset($_SESSION['nombre1'])) {
+  header('Location:iniciar_sesion.php?error=2');
+  echo "no esta iniciando la sesion";
+} elseif (isset($_SESSION['nombre1'])) {
+  require '../Dao/usuariosDao.php';
+  require '../Dto/usuariosDto.php';
+  require '../Dao/tiendaDao.php';
+  require '../Dto/tiendaDto.php';
+  if (isset($_SESSION["rol_id_Rol"])) {
+    $rol_id_Rol = $_SESSION["rol_id_Rol"];
+  }
+  $nombreTienda = $_SESSION["nombreTienda"];
+  $codigo_invitacion = $_SESSION["codigo_invitacion"];
+} else {
+  echo 'ocurrio un error';
+}
+echo "hola";
+var_dump($_POST);
+if (isset($_POST['registrarProducto'])) {
     $pDao = new productoDao();
     $pDto = new productoDto();
-    $pDto->setNombre( $_POST['Nombre']);
+    $pDto->setNombre($_POST['Nombre']);
     $pDto->setPrecio_unit($_POST['Precio_unit']);
     $pDto->setDescripción($_POST['Descripción']);
     $pDto->setMarca($_POST['Marca']);
@@ -21,11 +39,29 @@ if (isset($_POST['registrarProducto'])){
     if ($mensaje === 'Registrado Exitosamente') {
         // Registration successful, redirect to login page or success page
         header("Location:../../PAGINA/registro.php?registro=exitoso");
-        exit;
+        exit();
     }
-
-}
-else if (isset($_POST['registrarProducto'])){
+} else if (isset($_POST['registrarProductoUnico'])) {
+    echo "hola";
+    $pDao = new productoDao();
+    $pDto = new productoDto();
+    $pDto->setNombre($_POST['Nombre']);
+    $pDto->setPrecio_unit($_POST['Precio_unit']);
+    $pDto->setDescripción('');
+    $pDto->setMarca('');
+    $pDto->setCategoría('');
+    $pDto->setPresentacion('');
+    $pDto->setFecha_vencimiento('');
+    $pDto->setStock($_POST['Stock']);
+    $pDto->setStock_Min('');
+    $mensaje = $pDao->registrarProductoUnico($pDto, $_SESSION['codigo_invitacion']);
+    echo $mensaje;
+    if ($mensaje === 'Registrado Exitosamente') {
+        // Registration successful, redirect to login page or success page
+        header("Location:../../PAGINA/gestionar_productos.php?registro=exitoso");
+        exit();
+    }
+} else if (isset($_POST['registrarProducto'])) {
     $pDao = new productoDao();
     $pDto = new productoDto();
     $pDto->setId_Producto($_POST['id_Producto']);
@@ -44,16 +80,16 @@ else if (isset($_POST['registrarProducto'])){
     echo $mensaje;
     if ($mensaje === 'Registrado Exitosamente') {
         header("Location:../tablas/producto/listarproducto.php?mensaje=registro exitoso");
-        exit;
+        exit();
     }
-}
-else if ($_GET['id_Producto']!=null){
-    $pDao = new productoDao();
-    $mensaje = $pDao->eliminarProducto($_GET['id_Producto']);
-    header("Location:../tablas/producto/listarproducto.php?mensaje=".$mensaje);
-    exit();
-}
-else if (isset($_POST['modificarProducto'])){
+} else if (isset($_GET['id_Producto'])) {
+    if ($_GET['id_Producto'] != null) {
+        $pDao = new productoDao();
+        $mensaje = $pDao->eliminarProducto($_GET['id_Producto']);
+        header("Location:../tablas/producto/listarproducto.php?mensaje=" . $mensaje);
+        exit();
+    }
+} else if (isset($_POST['modificarProducto'])) {
     $pDao = new productoDao();
     $pDto = new productoDto();
     $pDto->setId_Producto($_POST['id_Producto']);
@@ -68,6 +104,6 @@ else if (isset($_POST['modificarProducto'])){
     $pDto->setStock_Min($_POST['Stock_Min']);
     $pDto->setinventario_id_Inventario($_POST['inventario_id_Inventario']);
 
-    $mensaje =$pDao->modificarProducto($pDto);
-    header("Location:../tablas/producto/listarproducto.php?mensaje=".$mensaje);
+    $mensaje = $pDao->modificarProducto($pDto);
+    header("Location:../tablas/producto/listarproducto.php?mensaje=" . $mensaje);
 }
